requirements.txt
opencv-python-headless
streamlit
streamlit-webrtc
deepface
numpy
retina-face
faiss-cpu